Treat registers as signed in branch instructions with comparisons
* The documentation doesn't state this, but it's necessary. On the loop
at 0xbfc43140 the register s3 is loaded with 0x27 and is used as the counter
in a for loop. However because its value wasn't treated as signed so it looked
more like 0xffffffffffffff27, which is very large, making the loop run forever.
Fix this by treating registers as signed where needed