Wrap log funtion into a macro and implement BIOS console
* Since log output is getting very large, it's common to have to wait
5+ minutes before any unknow instruction is encoutered. Printing to the
console actually takes a lot of time and slows down interpretation
significantly. Right now we don't care, since we just want to boot the BIOS
but let's have an option to disable all the log messages if we want.

* In addition record all writes to 0xb000f180 which is the BIOS console
output address, so we can have some output, which will be very useful