Ignore unrecognized reads/writes
* Currently the BIOS only writes to scratchpad and some very few mysterious
addresses that don't seem to do anything. However it is important to know
when it will try to write to DMAC for example so we can implement it. So
instead of writing anything and uncontrollably into a single large buffer
let's make an if-else with all the known addresses and how to handle them.
When the BIOS tries to write somewhere new we will be notified immediately.

* Also rework the disassembly logger to use C FILE* since these are faster
then std::ofstream. Normally I wouldn't care about this but in our usecase
which is very performance sensitive, it makes a noticeable difference.