Respect type of handler
* Until now the memory system didn't take into account the bit width
of the data coming in and out of the handlers. Instead I assumed that
64bits would be large enough for everything. But alas I was wrong.
Some addresses (notably the GIF/IPU FIFOs) are read/written with 128bit values.
I don't want to force every function to return __int128 types as that
would cripple performance so some tweaks were needed.

This isn't as hard as it might sound. The emulator read/write functions
are templates so we know which type we want beforehand. So it's as
simple as abstracting the Handler with a bit more inheritence magic and
we can cast HandlerBase to the type we want.