Add initial implementation of EE timers
* Yeah, timers again, my favourite topic... To be frank the EE timers
are a bit simpler than the IOP timers as they have less complexity
in their configuration. However, since the BIOS starts to use them
at this point, we can't get away with a extermelly partial implementation
like the IOP.

* The Emotion Engine has four hardware timers, each of them having
three registers (four on Timer 0 and 1). They are practically the same
with the IOP in that regard, having a count a compare/target and a mode
register. Timer 0 and 1 have an additional register Tn_HOLD which
keeps track of the count value when a peripheral on the
SBUS generates an interrupt.

* All the timers increment based on the bus clock which is exactly
half of the EE clock. The timers can also be configured to count
based on external sources, namely hblank and vblank. These are less
accurate but can be used to keep track when the screen refreshes.
I had hoped that we could have ignored hblank for now, but the BIOS
configures Timer 3 (used for BIOS alarms) to use it so implementing it
is necessary. The timings were taken from the timer header [1]
of the ps2sdk.

* An interesting fact as well is the interrupts as edge triggered
which means that an interrupt is sent to the EE when the raised flags in
Tn_MODE switch from 0 to 1 [2]. This is easy to implement and so did I,
to avoid any headaches in the future.

* Since the EE ticks the timers directly, we can't increment the counters
each time the function get called. To properly emulate the timer frequency,
an internal counter is used, that when its value is equal to the ratio
between the EE frequency and the timer clock, the real counter is incremented.

* This can be expensive since the timer function gets called every EE cycle
so we will probably change it to cycle adding in the future, especially when
the JIT will be implemented.

[1] https://github.com/ps2dev/ps2sdk/blob/master/ee/kernel/include/timer.h#L53
[2] https://psi-rockin.github.io/ps2tek/#eetimers