Major bug fixes + EE <-> IOP communication support
* This commit was intended to start the DMA implementation, but
quickly turned to a tedious bug fixing journey. So in return for
having my soul drained from all the debugging,
the following bugs were fixed:

* Fixed bug that caused incorrect IOP and EE timer writes
* Fixed bug that caused interrupts to not trigger
(NOTE: This is an error on ps2tek and more specifically with the
use of the I_CTRL register)
* Fixed handling of some virtual addresses (0x2*)
(The program doesn't notify me when out of bounds writes
happen for some reason?)
* Fixed handling of DICR2 register in the IOP DMA and
* Fixed INT1 interrupts running endlessly
* Fixed incorrect int1_pending field position in COP0 status

In addition to the above bugfixes some new additions needed to be
made to make DMA work:

* Added VBLANK ON/OFF interrupts
* Added INT1 interrupt detection
* Added some new EE instructions related to exception handling.
* Added syscall support

So after all of that, at least the program started running normally
again and I could begin with the DMA implementation. SIF0/SIF1 are
quite confusing since nowhere in ps2tek does it mention where that data
comes from. Thankfully I came across this linux kernel commit [1] which
specified that there exists a bidirection FIFO in the SIF that sends
data from either side to the other.

There are problems however. The EE DMAC works exclusively with qwords
or 4 word packets, while the IOP DMA only sends words. Also the DMAC
tags are 128bit while the IOP tags are 64bit. So even though I would
have liked to use qwords in the fifo, I was forced to switch to 32bit
values to have finer control on how much data gets used.

Other things to note are that the EE/IOP can start DMA transfer
without the FIFO having any data. In that case the transfer stalls
and waits until the other side starts filling it. A typical sequence
goes as follows:

1. The EE starts a SIF0 transfer, waiting for the IOP to send data.
2. The IOP starts a SIF0 transfer, pushing data from its RAM to the fifo.
3. The EE notices and starts taking that data to form a DMA tag
(when the fifo reaches a size bigger than 4 words of course)
5. The transfer completes and INT1 is asserted.
6. An exception is triggered and the exception handler checks the
data that was written.

If any of those steps above go wrong (IOP sends too much data, EE receives
it too early, the data is not written to the correct address) the whole
process will be stuck in an infinite loop. Now figure out what happened
in between the 1000+ instructions that were executed...
At least it seems to be working quite well now.

Note that this is not the final implementation of the DMA. Normally only
one channel can trasfer data in a single cycle while in our current
implementation all the channels are checked. I have a request system
in mind, that should fix this but right now I really want some graphics
on screen as fast as possible.

[1] https://patchwork.kernel.org/project/linux-mips/patch/fb79dab2db2bfa9a06e96c211d27423d0c51399c.1567326213.git.noring@nocrew.org/