GS: Add PSMCT32 texture support
* We are getting texture writes, exciting times. The BIOS initializes
a 256 x 64 PSMCT32 texture. The DMAC sends data with the GIF PATH3
to the the GS. So let me explain how texture writes work, cause the GS
is very peculiar in this regard.

* The VRAM in the GS is 4MB in size and is split into 8KB pages,
each of which is a grid of 256 byte blocks. Each block is divided
into 4 columns and finally each column contains the actual pixel data.
Here is a small little diagram:

-page
  - block
    - column
      - pixel
      - pixel
      ...
    - column
      - pixel
      - pixel
      ...
  - block
    - column
      - pixel
      - pixel
      ...

* The problem is that pixels and blocks aren't actually stored sequentially in
video ram. How they are ordered depends on the format, but here is an
example for PSMCT32:

 	<------- 8 blocks/64 pixels ------->
  | 0| | 1| | 4| | 5| |16| |17| |20| |21| ^
  | 2| | 3| | 6| | 7| |18| |19| |22| |23| | 4 blocks/32 pixels
  | 8| | 9| |12| |13| |24| |25| |28| |29| |
  |10| |11| |14| |15| |26| |27| |30| |31| |

* Say you want to write to the block at coordinates (5, 2) in the above page.
If the blocks were sequentially stored, then the offset of the block in memory would be:

x + width * y => 5 + 8 * 2 = 21

* In reality though block 25 would be accesed. Personally I don't
know why the GS organizes VRAM this way, but it makes emudevs like me
struggle a lot. I guess it has something to do with texture swizzling
but I hope I will know more in the short future. Need to also start
thinking if it is possible to read textures like this from shaders.
I guess it could be possible.

* Other misc changes in this build include:
- Unaligned stores/loads were rewritten to be branchless
- Fixed bug in GIF PATH3 that caused some packets to be ignored
- Switched to explicit register writing in the GS. This is
  because most registers are used to perform special operations,
  rather than just store data and writing structs is very inconvenient
  in this case. Will probably switch other components to this like
  (DMAC...)

Sources:
[1] https://tcrf.net/User:Kojin/TIM2_Information
[2] GS User's Manual Version 6.0