VIF: Refactor code and implement more functions
* This is a really big commit as it expands the previous VIF
implementation significantly. The VIF acts as a gateway between
the EE and the VU, letting the former transfer both instructions
and data to the latter.
Each VIFcode consists of a 16bit IMMEDIATE field, the interpretation
of which depends on the command, an 8bit NUM field that also changes
depending on the command and the 8bit CMD field.

* Interestingly the VIF considers instructions and data different
entities and handles them differently, with different instructions.
MPG is used to transfer instructions while UNPACK and its variants
transfer data exclusively. The NUM field in the MPG instruction
dictates the number of dwords transfered to the VU, while in UNPACK
NUM states the number of qwords WRITTEN to the VU. This is important
because the number or read and write data in UNPACK is not necessarily equal.

* UNPACK basically works by reading some amount of words from the FIFO
and converting them into a qword. Some formats are straightforward like
V4-32, which is implemented here, that reads 4 words and packs them into
a qword. Other formats are more complex. In addition though, UNPACK
has many many other variables to consider. Most important is the CYCLE
register that enables skipping or filling writes according to the
relation between the CL and WL fields. Skipping write is simple;
it basically means "write CL qwords and jump ahead (WL - CL) qwords".
Filling write is another story that I won't explain here, since it's
not needed at this point.

* In addition the STL queue has been replaced in favour of a custom written
leightweight Queue class. This desicion was made for several reasons:

1. std::queue is based on std::deque, which has overhead
2. Most STL containers in general are too "safe" for us.
In UNPACK for example, there are several formats that require
pulling multiple words of data from the FIFO at once. This can
be done with the STL queue, but is tedious.
3. STL queue is dynamic. All PS2 FIFOs are static, either 32 or 64 qwords
in size. When they fill up the DMA stalls until the component starts
draining the FIFO. This behaviour is crucial for some games and must
be emulated. Using a dynamic std::queue turns this into constant
size checks that hurt performance.

Sources
[1] https://psi-rockin.github.io/ps2tek/#vif
[2] EE User's Manual, chapter 6.3