EE: Implement missed LWR behaviour
* Was scrolling through the CPU manual again and noticed this
remark on the LWR instruction description:

"If the word sign bit (bit 31) is loaded from memory into the register by the instruction,
then the loaded word is sign-extended. If the sign bit is not loaded from memory by the
LWR, then bits 63..32 of the destination are unchanged", TX79 Core Architecture [A-75]

Before implementing this I checked if other emulators also handle this
and it seems the answer is positive [1]. So I added this, since
it's not too difficult and should prevent headaches in the future.

[1] https://github.com/PSI-Rockin/DobieStation/blob/master/src/core/ee/emotioninterpreter.cpp#L796