EE: Add test elf loading support
* Going this long without running any tests is like walking
on charcoal in emulation development. So I wrote a little
elf loader that loads the ELF file into RAM and executes
it, when the BIOS has fully loaded, because most tests
require some basic syscalls to be available.

* Using the EE tests the following bugs were fixed:

1. Fix branch delays in BEQ/JALR instructions
(NOTE: That test wasn't using valid MIPS code, but
it's nice to know we are accurate to what the hardware does)
2. DIV/DIVU support for dividing with zero
3. Added missing sign extension to ADDU