VU: Rework VU writes to be more generic
* Accessing the VU memory is more complex than it seems.
Both VUs have their memory mapped to the main bus between
0x11000000 - 0x11010000. However the distinction between
code and data once again comes to make this more complicated.
Inside the VU code and data are considered different memory
spaces, so their address spaces also different. This means
that an address of 0x0 can refer to either, depending on the
caller instruction.

* Using this ahead-of-time knowledge we can use templates to make
the compiler do the work for us and just have a small branch when
the EE wants to write something directly, which is pretty rare.