GS: [WIP] Handle depth properly
* Depth testing in the GS is controlled by the TEST_1/TEST_2
registers. The problem is that nothing prevents the program
from changing these register mid frame, meaning we have to
flush our renderer and rebuild the VBO from scratch.

* In addition enable OpenGL depth testing and configure it
properly based on the TEST configuration. It's also important
to adjust the depth buffer to have a default value of 0 rather
than 1, otherwise GL_GREATER tests will always fail.

* These changes make the triangles with proper depth. Need to also
implement texture transfers to complete the demo. This is more complex
than it sounds for multiple reasons that will hopefully be explained
in the following commits. Currently the GS code is extermely hacky
and unpleasant but I will clean it up soon, I promise.