EE: Start work on MIPS JIT compiler
* This is something that I wanted to do for a very long time but I haven't
due to lack of experience and for practical reasons since older systems
don't really see any performance benefits from JITs. This is not the
case with the PS2.

* The JIT will use the asmjit library for code generation as it handles
all the system calling conventions and allows us to focus on the code
generation part. For now the JIT is in the prototyping phase.

* The general architecture will be:
    IRBuilder -> Passes -> CodeGenerator
  I have some notes and I hope I can explain more things as I go.